requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
aiostream==0.5.2
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.5.14
demjson3==3.0.6
Deprecated==1.2.14
distro==1.8.0
docstring-parser==0.15
exceptiongroup==1.2.0
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.10.0
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.6
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.344
langchain-core==0.0.8
langsmith==0.0.68
llama-index==0.8.69.post2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
openai==1.3.7
packaging==23.2
pandas==2.1.3
peewee==3.17.0
prettytable==3.9.0
prompt-toolkit==3.0.36
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
pymemgpt==0.2.5
pypdf==3.17.1
python-box==7.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
questionary==2.0.1
regex==2023.10.3
requests==2.31.0
rich==13.7.0
safetensors==0.4.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.1
tokenizers==0.15.0
torch==2.1.1
tqdm==4.66.1
transformers==4.35.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
wcwidth==0.2.12
websockets==12.0
wrapt==1.16.0
yarl==1.9.3
